Move Article and CreateArticle to component
